Updated Prog
Added sig fig rounding for printing the equations for the fits. prints polynomials coefficients for the poly fit. still has odd, inconsistent errors/bugs when using the x:19 y:2 sample, still trying to figure that out. Either it stops the program due to an error or it does not plot the fi over the entire data set.