Update
Added inline graphs for easier comparison. still has issues with the unfiltered samples where it continues to print the polynomial coefficients in scientific notation even though for the order of magnitude that they are at it has been programmed to not print in scientific notation. these only occur for the x=19 y=2 trials and not for the x=21 y=2